refactor(storage): centralize localStorage
- Create shared storage service for avatar operations
- Move localStorage logic from components to service
- Update components to use centralized storage service
