+++ src/features/avatar/components/options-picker.tsx
@@ -1,9 +1,9 @@
 import '@/shared/styles/options-picker.css';
 import TabBar from '@/shared/components/ui/tab-bar';
 import { CustomizationOptions, TabData } from '@/features/avatar/types';
-import { customizationOptions } from '@/features/avatar/services';
+import { customizationOptions } from '@/features/avatar/services/avatar';
 import { useState } from 'react';
-import { buildURL } from '@/features/avatar/services';
+import { buildURL } from '@/features/avatar/services/avatar';
 import { useContext } from 'react';
 import { AvatarContext } from '@/features/avatar/context';
