+++ src/features/avatar/hooks.ts
@@ -1,17 +1,10 @@
-import { AvatarList } from '@/features/avatar/types';
+import { storageService } from '@/shared/services/storage';
 
 export const useOnUpdateAvatarList = () => {
   try {
-    const keys = Object.keys(window.localStorage);
-    const aList: AvatarList = [];
-    keys.forEach((key) => {
-      let item = window.localStorage.getItem(key) || "{URL:'undefined', name: 'undefined'}";
-      const avatar = JSON.parse(item);
-      avatar.key = key;
-      aList.push(avatar);
-    });
-    return aList;
+    return storageService.getAllAvatars();
   } catch (error) {
-    console.log(error);
+    console.error('Failed to get avatars:', error);
+    return [];
   }
 };
