+++ src/shared/components/ui/color-picker.tsx
@@ -1,7 +1,7 @@
 import { useState } from 'react';
 import { useContext } from 'react';
 import { AvatarContext } from '@/features/avatar/context';
-import { COLOR_PALETTE } from '@/features/avatar/services';
+import { COLOR_PALETTE } from '@/features/avatar/services/avatar';
 import '@/shared/styles/ui/color-picker.css';
 
 interface Props {
